+++ Assets/Project/Modules/Database/Scripts/Weapon/IWeaponConfig.cs
@@ -3,7 +3,7 @@ namespace Game.Database
     public interface IWeaponConfig
     {
         WeaponType Type { get; }
-        float Damage { get; }
+        int Damage { get; }
         float ShootInterval { get; }
         float Range { get; }
         SFXTypeProjectile SfxOnShoot { get; }
