+++ Assets/Project/Modules/Database/Scripts/Weapon/WeaponConfig.cs
@@ -7,7 +7,7 @@ namespace Game.Database
     public class WeaponConfig : IWeaponConfig
     {
         [field: SerializeField] public WeaponType Type { get; set; }
-        [field: SerializeField] public float Damage { get; set; } = 1f;
+        [field: SerializeField] public int Damage { get; set; } = 1;
         [field: SerializeField] public float ShootInterval { get; set; } = 0.3f;
         [field: SerializeField] public float Range { get; set; } = 10f;
         [field: SerializeField] public SFXTypeProjectile SfxOnShoot { get; set; }
