+++ Assets/Project/Modules/Game/Scripts/Weapon/DamageDealer.cs
@@ -1,11 +1,10 @@
-using Game.Database;
 using UnityEngine;
 
 namespace Game
 {
     public class DamageDealer : MonoBehaviour
     {
-        [field: SerializeField] public float Damage { get; private set; } = 1f;
+        [field: SerializeField] public int Damage { get; private set; } = 1;
 
         public float ID { get; private set; }
 
@@ -14,7 +13,7 @@ namespace Game
             this.ID = Random.value;
         }
 
-        public void SetDamage(float damage)
+        public void SetDamage(int damage)
         {
             this.Damage = damage;
         }
